Remove duplicate code from asm-x64.h
The structure of the X64 assembler changes as follows: asm-x64-intelxed.h
contains the XED-based X64Assembler class, asm-x64-legacy.h contains
the previous version of the assembler and asm-x64.h contains the code
commonly used by the two. Struct MemoryRef has a specialized
implementation for XED, so its implementation is moved from asm-x64.h
and now both asm-x64-intelxed.h and asm-x64-legacy.h include their own
specific version of this struct. asm-x64.h includes either
asm-x64-intelxed.h or asm-x64-legacy.h depending on the status of
HAVE_LIBXED (defined/not defined).
